refactor(web): split reconnect recovery helper
